feat(sdk-coin-bch): refactor bch to its own module
- Move bch, tbch and their tests to its own module.

Ticket: BG-51031
