+++ modules/bitgo/src/v2/coins/bsv.ts
@@ -5,7 +5,7 @@ import { BitGoBase, BaseCoin } from '@bitgo/sdk-core';
 import * as utxolib from '@bitgo/utxo-lib';
 
 import { UtxoNetwork } from '@bitgo/abstract-utxo';
-import { Bch } from './bch';
+import { Bch } from '@bitgo/sdk-coin-bch';
 
 export class Bsv extends Bch {
   constructor(bitgo: BitGoBase, network?: UtxoNetwork) {
