+++ modules/bitgo/src/v2/coins/index.ts
@@ -4,6 +4,8 @@ import { Algo, AlgoToken, Talgo } from '@bitgo/sdk-coin-algo';
 export { Algo, AlgoToken, Talgo };
 import { AvaxP, TavaxP } from '@bitgo/sdk-coin-avaxp';
 export { AvaxP, TavaxP };
+import { Bch, Tbch } from '@bitgo/sdk-coin-bch';
+export { Bch, Tbch };
 import { Btc, Tbtc } from '@bitgo/sdk-coin-btc';
 export { Btc, Tbtc };
 import { Cspr, Tcspr } from '@bitgo/sdk-coin-cspr';
@@ -15,7 +17,6 @@ export { Xrp, Txrp };
 import { Zec, Tzec } from '@bitgo/sdk-coin-zec';
 export { Zec, Tzec };
 
-export { Bch } from './bch';
 export { Bsv } from './bsv';
 export { Btg } from './btg';
 export { Celo } from './celo';
@@ -34,7 +35,6 @@ export { StellarToken } from './stellarToken';
 export { Susd } from './susd';
 export { FiatUsd } from './fiatusd';
 export { FiatEur } from './fiateur';
-export { Tbch } from './tbch';
 export { Tbsv } from './tbsv';
 export { Tcelo } from './tcelo';
 export { Teos } from './teos';
