+++ modules/sdk-coin-bch/.eslintignore
@@ -1,0 +1,4 @@
+node_modules
+.idea
+public
+dist
