+++ modules/sdk-coin-bch/.prettierignore
@@ -1,0 +1,3 @@
+.nyc_output/
+dist/
+mochawesome-report/
