+++ modules/sdk-coin-bch/src/bch.ts
@@ -3,7 +3,6 @@ import { BitGoBase, BaseCoin } from '@bitgo/sdk-core';
 import * as utxolib from '@bitgo/utxo-lib';
 
 export class Bch extends AbstractUtxoCoin {
-
   protected constructor(bitgo: BitGoBase, network?: UtxoNetwork) {
     super(bitgo, network || utxolib.networks.bitcoincash);
   }
