+++ modules/sdk-coin-bch/src/index.ts
@@ -1,0 +1,2 @@
+export * from './bch';
+export * from './tbch';
