+++ modules/sdk-coin-bch/src/tbch.ts
