+++ tsconfig.packages.json
@@ -29,6 +29,9 @@
       "path": "./modules/sdk-coin-btc"
     },
     {
+      "path": "./modules/sdk-coin-bch"
+    },
+    {
       "path": "./modules/sdk-coin-cspr"
     },
     {
